fix gamma handling for GigE cameras
When connecting to a GigE camera (aca1920-50gc), the node crashes with "Error while accessing Gamma in PylonCameraImpl<CameraTraitT>": For some camera types, the Gamma settings are not available to the interface as they are handled automatically by the camera itself. This was already partly fixed for some gamma-related function calls, but not for all of them. This fix adds the missing checks. The behaviour is slightly changed: If the gamma is not set via user, gamma remains in auto mode (i.e. controlled by the camera). Once the user tries to set a gamma value, the gamma mode switches to user.